refactor: Change location of JSONValueFactory in a separate file
