+++ JSONParser/JSONParser/Parser.swift
@@ -41,22 +41,3 @@ struct Parser {
     }
     
 }
-
-struct JSONValueFactory {
-    static let trueString = "true"
-    static let falseString = "false"
-    
-    static func make(token: String) -> JSONValue {
-        if token.first == JSONSymbols.doubleQuotation {
-            return String(token)
-        }
-        if token == trueString {
-            return true
-        }
-        if token == falseString {
-            return false
-        }
-        return Int(token)!
-    }
-    
-}
